refactoring(makin sure it works, cleaning impot and code)
