+++ bionc/utils/c3d_ik_exporter.py
@@ -1,26 +1,35 @@
+from math import ceil, log10
+
 import ezc3d
 import numpy as np
-from bionc.bionc_numpy import NaturalCoordinates
-from math import ceil, floor, log10
-from export_c3d_from_bionc_model import add_point_from_dictionary
+
+from .c3d_export_utils import add_point_from_dictionary
+from ..bionc_numpy import NaturalCoordinates
+from ..protocols.biomechanical_model import GenericBiomechanicalModel
 
 
 class C3DInverseKinematicsExporter:
-    def __init__(self, filename, model):
+    def __init__(self, filename: str, model: GenericBiomechanicalModel):
+        """
+        Parameters
+        ----------
+        filename: str
+            The file of the original c3d to consider
+        model: GenericBiomechanicalModel
+            The Biomechanical model considered
+        """
+
         self.filename = filename
         self.c3d_file = ezc3d.c3d(filename)
         self.model = model
 
-    def add_technical_markers_to_c3d(self, Q) -> None:
+    def add_technical_markers(self, Q: NaturalCoordinates) -> None:
         """
         This function add the technical markers of the model to the c3d file. This point are the markers that are rigidly associated to the
         segments of the model.
+
         Parameters
         ----------
-        c3d_file : ezc3d.c3d
-            The c3d file to add the points to
-        model : BiomechanicalModel
-            The biomechanical model from which the data can be exported
         Q : numpy.ndarray | NaturalCoordinates
             The natural coordinates of the model, should be given as a Natural coordinates but if given as a numpy array it will be converted
 
@@ -42,15 +51,12 @@ class C3DInverseKinematicsExporter:
 
         add_point_from_dictionary(self.c3d_file, dict_to_add)
 
-    def add_natural_coordinate_to_c3d(self, Q) -> None:
+    def add_natural_coordinate(self, Q: NaturalCoordinates) -> None:
         """
         This function add the natural coordinate of the model to the c3d file. It add the segment rp,rd,u,w to the c3d file.
+
         Parameters
         ----------
-        c3d_file : ezc3d.c3d
-            The c3d file to add the points to
-        model : BiomechanicalModel
-            The biomechanical model from which the data can be exported
         Q : numpy.ndarray | NaturalCoordinates
             The natural coordinates of the model, should be given as a Natural coordinates but if given as a numpy array it will be converted
 
@@ -65,7 +71,6 @@ class C3DInverseKinematicsExporter:
         # Calulation of a reasonable factor for the u and w
         list_factor = []
         for s in range(Q.nb_qi()):
-            name_segment = self.model.segment_names[s]
             Qi = Q.vector(s)
             rp_temp = Qi.rp
             rd_temp = Qi.rd
@@ -87,8 +92,6 @@ class C3DInverseKinematicsExporter:
             u_temp = Qi.u
             w_temp = Qi.w
 
-            v_mean = np.mean(np.linalg.norm(rd_temp - rp_temp, axis=0))
-            u_mean = np.mean(np.linalg.norm(u_temp, axis=0))
             dict_to_add[f"u_{name_segment}"] = rp_temp + u_temp / factor
             dict_to_add[f"rp_{name_segment}"] = rp_temp
             dict_to_add[f"rd_{name_segment}"] = rd_temp
